Refactor types to be more modular
